Adds end-to-end generation governance: per-run audit tables, per-claim
verdicts, drift scoring, and quality persistence through the fallback path.

- governance_runs and claim_verdicts tables with RLS
- jobs columns: governance_passed, governance_drift_score, last_governance_run_id
- responsibilities included in evidence so copied bullets are not falsely flagged
- fallback generation path now persists governance/quality alongside primary path

E2E verified on Vercel preview: governance_passed=true, drift_score=0,
15 claim verdicts all high-confidence, quality_passed=true.

+++ scripts/017_create_job_scores_table.sql
@@ -3,7 +3,7 @@
 
 CREATE TABLE IF NOT EXISTS public.job_scores (
   job_id uuid PRIMARY KEY REFERENCES public.jobs(id) ON DELETE CASCADE,
-  overall_score integer NOT NULL,
+  overall_score numeric(5,2) NOT NULL,
   skills_match numeric(5,2) NOT NULL,
   experience_relevance numeric(5,2) NOT NULL,
   evidence_quality numeric(5,2) NOT NULL,
